+++ src/main/java/com/example/gruppe4bilabonnement/repositories/MechanicRepository.java
@@ -23,20 +23,20 @@ public class MechanicRepository {
         String query = "SELECT * FROM damage_report;";
         return jdbcTemplate.query(query, (rs, rowNum) -> {
             DamageReport damageReport = new DamageReport();
-            damageReport.setId(rs.getLong("id"));
-            damageReport.setCarId(rs.getLong("car_id"));
+            damageReport.setId(rs.getInt("id"));
+            damageReport.setCarId(rs.getInt("car_id"));
             damageReport.setDamageText(rs.getString("damage_text"));
             damageReport.setPrice(rs.getDouble("price"));
             return damageReport;
         });
     }
 
-    public DamageReport getDamageReportById(Long id) {
+    public DamageReport getDamageReportById(int id) {
         String query = "SELECT * FROM damage_report WHERE id = ?;";
         return jdbcTemplate.queryForObject(query, new Object[]{id}, (rs, rowNum) -> {
             DamageReport damageReport = new DamageReport();
-            damageReport.setId(rs.getLong("id"));
-            damageReport.setCarId(rs.getLong("car_id"));
+            damageReport.setId(rs.getInt("id"));
+            damageReport.setCarId(rs.getInt("car_id"));
             damageReport.setDamageText(rs.getString("damage_text"));
             damageReport.setPrice(rs.getDouble("price"));
             return damageReport;
@@ -48,12 +48,12 @@ public class MechanicRepository {
         jdbcTemplate.update(query, damageReport.getCarId(), damageReport.getDamageText(), damageReport.getPrice(), damageReport.getId());
     }
 
-    public void deleteDamageReport(Long id) {
+    public void deleteDamageReport(int id) {
         String query = "DELETE FROM damage_report WHERE id = ?;";
         jdbcTemplate.update(query, id);
     }
 
-    public List<DamageReport> getAllDamageReportsByCarId(Long carId) {
+    public List<DamageReport> getAllDamageReportsByCarId(int carId) {
         String query = "SELECT * FROM damage_report WHERE car_id = ?;";
         return jdbcTemplate.query(query, new DamageReportRowMapper(), carId);
     }
