+++ src/main/java/com/example/gruppe4bilabonnement/services/MechanicService.java
@@ -2,7 +2,6 @@ package com.example.gruppe4bilabonnement.services;
 
 import com.example.gruppe4bilabonnement.models.DamageReport;
 import com.example.gruppe4bilabonnement.repositories.MechanicRepository;
-import com.example.gruppe4bilabonnement.services.rowmappers.DamageReportRowMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,11 +21,11 @@ public class MechanicService {
         return mechanicRepository.getAllDamageReports();
     }
 
-    public DamageReport getDamageReportById(Long id) {
+    public DamageReport getDamageReportById(int id) {
         return mechanicRepository.getDamageReportById(id);
     }
 
-    public List<DamageReport> getAllDamageReportsByCarId(Long carId) {
+    public List<DamageReport> getAllDamageReportsByCarId(int carId) {
         return mechanicRepository.getAllDamageReportsByCarId(carId);
     }
 
@@ -34,7 +33,7 @@ public class MechanicService {
         mechanicRepository.updateDamageReport(damageReport);
     }
 
-    public void deleteDamageReport(Long id) {
+    public void deleteDamageReport(int id) {
         mechanicRepository.deleteDamageReport(id);
     }
 }
